+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Fixed
   * AZ-251: Align velero 1.4.0 to main branch of velero-plugin-for-microsoft-azure to work with managed identities
   * AZ-251: Remove hack for velero pod labels since it's now supported in chart v2.12.13
   * AZ-252: Update AGIC helm chart to 1.2.0 final
+  * AZ-254: Application gateway creation fail with default parameters.
   
 Added
   * AZ-253: Update to last stable version of AKS by default.